backend/theming: backport theming/preset module changes from #679
