+++ gradience/backend/globals.py
@@ -23,15 +23,14 @@ from gi.repository import Xdp
 from gradience.backend import constants
 
 
-presets_dir = os.path.join(
-    os.environ.get("XDG_CONFIG_HOME", os.environ["HOME"] + "/.config"),
-    "presets"
+user_config_dir = os.environ.get(
+    "XDG_CONFIG_HOME", os.environ["HOME"] + "/.config"
 )
 
-preset_repos = {
-    "Official": "https://github.com/GradienceTeam/Community/raw/next/official.json",
-    "Curated": "https://github.com/GradienceTeam/Community/raw/next/curated.json"
-}
+presets_dir = os.path.join(
+    user_config_dir,
+    "presets"
+)
 
 user_plugin_dir = os.path.join(
     os.environ.get("XDG_DATA_HOME", os.environ["HOME"] + "/.local/share"),
@@ -44,19 +43,50 @@ system_plugin_dir = os.path.join(
     "plugins"
 )
 
-def get_gtk_theme_dir(app_type):
+preset_repos = {
+    "Official": "https://github.com/GradienceTeam/Community/raw/next/official.json",
+    "Curated": "https://github.com/GradienceTeam/Community/raw/next/curated.json"
+}
+
+# Adwaita named UI colors prefixes list
+# NOTE: Remember to update this list if new libadwaita version brings up new variables
+adw_variables_prefixes = [
+    "accent_",
+    "destructive_",
+    "success_",
+    "warning_",
+    "error_",
+    "window_",
+    "view_",
+    "headerbar_",
+    "card_",
+    "dialog_",
+    "popover_",
+    "shade_",
+    "scrollbar_",
+    "borders"
+]
+
+# Adwaita named palette colors prefixes list
+# NOTE: Remember to update this list if new libadwaita version brings up new variables
+adw_palette_prefixes = [
+    "blue_",
+    "green_",
+    "yellow_",
+    "orange_",
+    "red_",
+    "purple_",
+    "brown_",
+    "light_",
+    "dark_"
+]
+
+def get_gtk_theme_dir(app_type: str):
     if app_type == "gtk4":
-        theme_dir = os.path.join(
-            os.environ.get("XDG_CONFIG_HOME",
-                            os.environ["HOME"] + "/.config"),
-            "gtk-4.0"
-        )
-    elif app_type == "gtk3":
-        theme_dir = os.path.join(
-            os.environ.get("XDG_CONFIG_HOME",
-                            os.environ["HOME"] + "/.config"),
-            "gtk-3.0"
-        )
+        theme_dir = os.path.join(user_config_dir, "gtk-4.0")
+
+    if app_type == "gtk3":
+        theme_dir = os.path.join(user_config_dir, "gtk-3.0")
 
     return theme_dir
 
@@ -66,6 +96,3 @@ def is_sandboxed():
     is_sandboxed = portal.running_under_sandbox()
 
     return is_sandboxed
-
-def get_available_sassc():
-    pass
