+++ gradience/backend/theming/meson.build
@@ -3,6 +3,6 @@ themingdir = 'gradience/backend/theming'
 gradience_sources = [
     '__init__.py',
     'monet.py',
-    'preset_utils.py'
+    'preset.py'
 ]
 PY_INSTALLDIR.install_sources(gradience_sources, subdir: themingdir)
