+++ gradience/backend/utils/meson.build
@@ -3,6 +3,7 @@ utilsdir = 'gradience/backend/utils'
 gradience_sources = [
     '__init__.py',
     'colors.py',
-    'common.py'
+    'common.py',
+    'theming.py'
 ]
 PY_INSTALLDIR.install_sources(gradience_sources, subdir: utilsdir)
